Adding options to mkdeps to either run the download or build steps only. Default is to do both. The reason for this change is that on Stellar we need to build the whole tool-chain, including dependencies, on the compute node (which is AMD for GPU-nodes and Intel for CPU-nodes). So the steps to build for Stellar GPUs is:
- Run mkdeps as:
  ./mkdeps.sh --build=no --build-openblas=yes --build-superlu=yes

- Then get a GPU node using the interactive and then run
  ./mkdeps.sh --download=no --build-openblas=yes --build-superlu=yes

Then build GkeyllZero itself on the GPU node you obtained.

Note: unfortunately, you need to have different builds for AMD and
Intel nodes. You can install in different locations using the --prefix
flag to the mkdeps and the configure scripts.

+++ install-deps/build-openblas.sh
@@ -10,17 +10,25 @@ DEP_SOURCES=$HOME/gkylsoft/dep_src/
 mkdir -p $DEP_SOURCES
 cd $DEP_SOURCES
 
-# delete old checkout and builds
-rm -rf OpenBLAS-*
+if [ "$DOWNLOAD_PKGS" = "yes" ]
+then
+    echo "Downloading OpenBLAS .."
+    # delete old checkout and builds
+    rm -rf OpenBLAS-*
+    curl -L https://github.com/xianyi/OpenBLAS/releases/download/v0.3.15/OpenBLAS-0.3.15.tar.gz > OpenBLAS-0.3.15.tar.gz
+fi
 
-curl -L https://github.com/xianyi/OpenBLAS/releases/download/v0.3.15/OpenBLAS-0.3.15.tar.gz > OpenBLAS-0.3.15.tar.gz
-gunzip -f OpenBLAS-0.3.15.tar.gz
-tar xvf OpenBLAS-0.3.15.tar
-cd OpenBLAS-0.3.15
-make USE_OPENMP=0 NUM_THREADS=1 FC=$FC -j
-make install PREFIX=$PREFIX -j
+if [ "$BUILD_PKGS" = "yes" ]
+then
+    echo "Building OpenBLAS .."
+    gunzip -f OpenBLAS-0.3.15.tar.gz
+    tar xvf OpenBLAS-0.3.15.tar
+    cd OpenBLAS-0.3.15
+    make USE_OPENMP=0 NUM_THREADS=1 FC=$FC -j
+    make install PREFIX=$PREFIX -j
 
-# soft-link 
-ln -sfn $PREFIX $GKYLSOFT/OpenBLAS
+    # soft-link 
+    ln -sfn $PREFIX $GKYLSOFT/OpenBLAS
+fi
